flytt NavAnsattRolle til annen pakke
