+++ mulighetsrommet-api/src/main/kotlin/no/nav/mulighetsrommet/api/navansatt/model/NavAnsattDto.kt
@@ -2,7 +2,6 @@ package no.nav.mulighetsrommet.api.navansatt.model
 
 import kotlinx.serialization.Serializable
 import no.nav.mulighetsrommet.api.clients.msgraph.AzureAdNavAnsatt
-import no.nav.mulighetsrommet.api.navansatt.db.NavAnsattRolle
 import no.nav.mulighetsrommet.model.NavIdent
 import no.nav.mulighetsrommet.serializers.LocalDateSerializer
 import no.nav.mulighetsrommet.serializers.UUIDSerializer
