+++ mulighetsrommet-api/src/main/kotlin/no/nav/mulighetsrommet/api/navansatt/model/NavAnsattRolle.kt
@@ -1,0 +1,11 @@
+package no.nav.mulighetsrommet.api.navansatt.model
+
+enum class NavAnsattRolle {
+    TEAM_MULIGHETSROMMET,
+    KONTAKTPERSON,
+    TILTAKSGJENNOMFORINGER_SKRIV,
+    AVTALER_SKRIV,
+    TILTAKADMINISTRASJON_GENERELL,
+    TILTAKADMINISTRASJON_ENDRINGSMELDING,
+    OKONOMI_BESLUTTER,
+}
